# Compile the dart_dev run script for better performance