Running a `dart_dev` command involves several subprocesses. The main reason for this is because we wanted to support rich package configuration via a `.dart` file that can do whatever is necessary and possible, rather than just a static config file in something like yaml. To make this work with `dart run dart_dev`, however, requires generating another Dart program that imports this config file from the current package and uses it when running `dart_dev`, which we then run via a subprocess.

The execution flow looks like this:

<img width="368" alt="Screen Shot 2021-11-20 at 11 04 23 AM" src="https://user-images.githubusercontent.com/1738457/142736734-b2f22613-dd55-4fef-bdf0-838fd7a40b7e.png">

With the exception that currently we don't perform the second precompilation step, meaning that every time you run `dart_dev`, you have to pay the cost of Dart compiling that generated run script. In some informal benchmarking, that step takes about **5 seconds**.

This PR introduces logic that will compile this run script and store the result in `.dart_tool/dart_dev/` so that subsequent runs can use it directly. Running precompiled Dart executables is very fast (~0.5s for kernel snapshots and ~0.02s for native executables), so once the compilation step is done, subsequent runs save almost all of the aforementioned 5 seconds.

The compiled result is cached along with a digest of its inputs. If any of these inputs change, the script will be recompiled:
- `.dart_tool/package_config.json` (pub get resolution + the current Dart SDK)
- `tool/dart_dev/config.dart`
- `tool/**.dart` (if the above config file has any relative imports)

Additionally, this compilation step will be skipped if:
- `tool/dart_dev/config.dart` imports from any package that is considered mutable

## Benchmarking

Here's some more informal benchmarking to demonstrate how this impacts first runs and subsequent (cached) runs:

### `dart run dart_dev` (first run)
| SDK                | Before | After  | Delta |
| ------------------ | ------ | ------ | ----- |
| 2.13.4             | ~15.6s | ~21s   | +5.4s |
| 2.14.4             | ~12.6s | ~18s   | +5.4s |
| 2.15.0-268.18.beta | ~11.7s | ~15.5s | +3.8s |

### `dart run dart_dev` (cached run)

| SDK                | Before | After | Delta |
| ------------------ | ------ | ----- | ----- |
| 2.13.4             | ~6.7s  | ~1.5s | -5.2s |
| 2.14.4             | ~7s    | ~2s   | -5s   |
| 2.15.0-268.18.beta | ~6.5s  | ~1.3s | -5.2s |

To summarize, the first run takes a perf hit of about 5 seconds, but as a result, subsequent runs have very little overhead and save that same 5 seconds every time.

---

This PR also adds a `clean` subcommand to `dart run dart_dev` that will remove the files in `.dart_tool/dart_dev/`, in case a bug is found in our compilation caching logic.