@alexandercampbell-wk 2024-03-05
I've revived this PR here with merge conflicts resolved and most comments addressed: #426
Because of adjustments to our CI and the number of intervening changes to the tool, it seemed cleaner to open a fresh PR.
I'll close this PR for now— if it's preferable to move that commit into this PR and reopen, I'd be happy to make that change.